增加用户信息功能

+++ Plugin.php
@@ -19,40 +19,18 @@ class WeChatHelper_Plugin implements Typecho_Plugin_Interface
      */
     public static function activate()
     {
-        $db = Typecho_Db::get();
-        if ("Pdo_Mysql" === $db->getAdapterName() || "Mysql" === $db->getAdapterName()) {
-            //创建自定义菜单表
-            $db->query("CREATE TABLE IF NOT EXISTS " . $db->getPrefix() . 'wch_menus' . " (
-				  `mid` int(11) NOT NULL AUTO_INCREMENT,
-				  `level` varchar(15) DEFAULT 'button',
-				  `name` varchar(32) DEFAULT '',
-				  `type` varchar(32) DEFAULT 'view',
-				  `value` varchar(200) DEFAULT '',
-				  `sort` int(3) DEFAULT '0',
-				  `order` int(3) DEFAULT '1',
-				  `parent` int(11) DEFAULT '0',
-				  `created` int(10) DEFAULT '0',
-				  PRIMARY KEY (`mid`)
-				) ENGINE=MyISAM  DEFAULT CHARSET=utf8 AUTO_INCREMENT=1");
-            $db->query("INSERT INTO `" . $db->getPrefix() . 'wch_menus' . "` (`mid`, `level`, `name`, `type`, `value`, `sort`, `order`, `parent`, `created`) VALUES
-				(1, 'button', '首页', 'view', 'https://www.iyuu.cn/', 10, 1, 0, 1503804104),
-				(3, 'button', '搜索', 'view', 'https://www.iyuu.cn/', 20, 2, 0, 1503804141),
-				(4, 'button', '其他', 'click', NULL, 50, 5, 0, 1503804153),
-				(6, 'sub_button', '最新文章', 'click', 'n', 51, 1, 4, 1503804247),
-				(7, 'sub_button', '随机文章', 'click', 'r', 52, 2, 4, 1503807202),
-				(8, 'sub_button', '手气不错', 'click', 'l', 54, 4, 4, 1503824995);");
-			//插入WCH_access_token、WCH_expires_in
-			$db->query($db->sql()->insert('table.options')->rows(array("name"=>"WCH_access_token","user"=>"0","value"=>"0")));
-			$db->query($db->sql()->insert('table.options')->rows(array("name"=>"WCH_expires_in","user"=>"0","value"=>"0")));
-        } else {
-            throw new Typecho_Plugin_Exception(_t('对不起, 本插件仅支持MySQL数据库。'));
-        }
-        
-		Helper::addRoute('wechat', '/wechat', 'WeChatHelper_Action', 'link');
-		Helper::addRoute('send', '/send', 'WeChatHelper_Widget_Send', 'send');
-		Helper::addAction('WeChat', 'WeChatHelper_Action');
-        Helper::addPanel(1, 'WeChatHelper/Page/Menus.php', '公众号菜单', '公众号菜单', 'administrator');
-        return ('微信助手已经成功激活，请进入设置Token!');
+        $msg = self::installDb();
+        //微信接口
+        Helper::addRoute('wechat', '/wechat', 'WeChatHelper_Action', 'link');
+        //模板消息发送接口
+        Helper::addRoute('send', '/[send:alpha].send', 'WeChatHelper_Widget_Send', 'send');
+        //后台管理接口
+        Helper::addAction('WeChat', 'WeChatHelper_Action');
+        //后台菜单
+        $index = Helper::addMenu('微信助手');
+        Helper::addPanel($index, 'WeChatHelper/Page/Users.php', '用户管理', '用户管理', 'administrator');
+        Helper::addPanel($index, 'WeChatHelper/Page/Menus.php', '自定义菜单', '自定义菜单', 'administrator');
+        return $msg;
     }
 
     /**
@@ -65,19 +43,85 @@ class WeChatHelper_Plugin implements Typecho_Plugin_Interface
      */
     public static function deactivate()
     {
-        $db = Typecho_Db::get();
-        $options = Typecho_Widget::widget('Widget_Options');
-        if ("Pdo_Mysql" === $db->getAdapterName() || "Mysql" === $db->getAdapterName()) {
-            $db->query("drop table " . $db->getPrefix() . "wch_menus");
-            $db->query($db->sql()->delete('table.options')->where('name like ?', "WCH_%"));
+        $dropTable = Typecho_Widget::Widget('Widget_Options')->plugin('WeChatHelper')->dropTable;
+        if (isset($dropTable) && $dropTable){
+            $db = Typecho_Db::get();
+            if ("Pdo_Mysql" === $db->getAdapterName() || "Mysql" === $db->getAdapterName()) {
+                $db->query("drop table " . $db->getPrefix() . "wch_menus, ".$db->getPrefix()."wch_users");
+                $db->query($db->sql()->delete('table.options')->where('name like ?', "WCH_%"));
+            }
         }
-
         Helper::removeRoute('wechat');
         Helper::removeRoute('send');
         Helper::removeAction('WeChat');
-        Helper::removePanel(1, 'WeChatHelper/Page/Menus.php');
+        $index = Helper::addMenu('微信助手');
+        Helper::removePanel($index, 'WeChatHelper/Page/Users.php');
+        Helper::removePanel($index, 'WeChatHelper/Page/Menus.php');
     }
-
+    /**
+	 * 安装数据库
+	 */
+	public static function installDb(){
+		try {
+            $db = Typecho_Db::get();
+            if ("Pdo_Mysql" === $db->getAdapterName() || "Mysql" === $db->getAdapterName()) {
+                //创建自定义菜单表
+                $db->query("CREATE TABLE IF NOT EXISTS " . $db->getPrefix() . 'wch_menus' . " (
+                    `mid` int(11) NOT NULL AUTO_INCREMENT,
+                    `level` varchar(15) DEFAULT 'button',
+                    `name` varchar(32) DEFAULT '',
+                    `type` varchar(32) DEFAULT 'view',
+                    `value` varchar(200) DEFAULT '',
+                    `sort` int(3) DEFAULT '0',
+                    `order` int(3) DEFAULT '1',
+                    `parent` int(11) DEFAULT '0',
+                    `created` int(10) DEFAULT '0',
+                    PRIMARY KEY (`mid`)
+                    ) ENGINE=MyISAM  DEFAULT CHARSET=utf8 AUTO_INCREMENT=1");
+                $db->query("INSERT INTO `" . $db->getPrefix() . 'wch_menus' . "` (`mid`, `level`, `name`, `type`, `value`, `sort`, `order`, `parent`, `created`) VALUES
+                    (1, 'button', '首页', 'view', 'https://www.iyuu.cn/', 10, 1, 0, 1503804104),
+                    (3, 'button', '最新文章', 'click', 'n', 20, 2, 0, 1503804141),
+                    (4, 'button', '其他', 'click', NULL, 30, 3, 0, 1503804153),
+                    (7, 'sub_button', '随机文章', 'click', 'r', 31, 1, 4, 1503807202),
+                    (8, 'sub_button', '手气不错', 'click', 'l', 32, 2, 4, 1503824995),
+                    (9, 'sub_button', '扫码', 'scancode_waitmsg', 'scan', 35, 5, 4, 1566277322),
+                    (12, 'sub_button', '地理位置', 'location_select', 'location_select', 33, 3, 4, 1566289962),
+                    (13, 'sub_button', '拍照相册', 'pic_photo_or_album', 'photo', 34, 4, 4, 1566290488);");
+                //创建用户管理表
+                $db->query("CREATE TABLE IF NOT EXISTS " . $db->getPrefix() . 'wch_users' . " (
+                        `uid` int(11) NOT NULL AUTO_INCREMENT,
+                        `openid` varchar(50) DEFAULT '',
+                        `nickname` varchar(100) DEFAULT '',
+                        `sex` char(1) DEFAULT '',
+                        `language` varchar(50) DEFAULT '',
+                        `city` varchar(50) DEFAULT '',
+                        `province` varchar(50) DEFAULT '',
+                        `country` varchar(50) DEFAULT '',
+                        `headimgurl` varchar(200) DEFAULT '',
+                        `subscribe_time` int(10) DEFAULT '0',
+                        `credits` int(10) NOT NULL DEFAULT '0',
+                        `bind` int(3) DEFAULT '0',
+                        `status` char(1) DEFAULT '1',
+                        `created` int(10) DEFAULT '0',
+                        `synctime` int(10) DEFAULT '0',
+                        PRIMARY KEY (`uid`)
+                        ) ENGINE=MyISAM  DEFAULT CHARSET=utf8 AUTO_INCREMENT=1");
+                //插入WCH_access_token、WCH_expires_in
+                $db->query($db->sql()->insert('table.options')->rows(array("name"=>"WCH_access_token","user"=>"0","value"=>"0")));
+                $db->query($db->sql()->insert('table.options')->rows(array("name"=>"WCH_expires_in","user"=>"0","value"=>"0")));
+            } else {
+                throw new Typecho_Plugin_Exception(_t('对不起, 本插件仅支持MySQL数据库。'));
+            }
+            return "数据表安装成功！";
+        } catch (Typecho_Db_Exception $e) {
+            if ('23000' == $e->getCode()) {
+                $msg = '数据表已存在!微信助手已经成功激活，请进入设置Token!';
+                return $msg;
+            }else{
+                return $e->getCode().'微信助手已经成功激活，请进入设置Token!';
+            }
+        }
+	}
     /**
      * 获取插件配置面板
      *
@@ -91,15 +135,15 @@ class WeChatHelper_Plugin implements Typecho_Plugin_Interface
         /** 用户添加订阅欢迎语 **/
         $welcome = new Typecho_Widget_Helper_Form_Element_Textarea('welcome', NULL, '欢迎！' . chr(10) . '发送\'h\'让小的给您介绍一下！', '订阅欢迎语', '用户订阅之后主动发送的一条欢迎语消息。');
         $form->addInput($welcome);
-        /** 返回最大结果条数 **/
-        $imageDefault = new Typecho_Widget_Helper_Form_Element_Text('imageDefault', NULL, 'https://holmesian.org/usr/themes/Holmesian/image/avatar.jpg', _t('默认显示图片'), '图片链接，支持JPG、PNG格式，推荐图为80*80。');
+        /** 图文默认图片 **/
+        $imageDefault = new Typecho_Widget_Helper_Form_Element_Text('imageDefault', NULL, 'https://cdn.iyuu.cn/usr/uploads/2019/08/1364235608.jpg', _t('默认显示图片'), '图片链接，支持JPG、PNG格式，推荐图为80*80。');
         $form->addInput($imageDefault);
         /** 返回最大结果条数 **/
-        $imageNum = new Typecho_Widget_Helper_Form_Element_Text('imageNum', NULL, '5', _t('返回图文数量'), '图文消息数量，限制为10条以内。');
+        $imageNum = new Typecho_Widget_Helper_Form_Element_Text('imageNum', NULL, '5', _t('返回图文数量'), '图文消息数量，限制为8条以内。');
         $imageNum->input->setAttribute('class', 'mini');
         $form->addInput($imageNum);
         /** 日志截取字数 **/
-        $subMaxNum = new Typecho_Widget_Helper_Form_Element_Text('subMaxNum', NULL, '200', _t('日志截取字数'), '显示单条日志时，截取日志内容字数。');
+        $subMaxNum = new Typecho_Widget_Helper_Form_Element_Text('subMaxNum', NULL, '50', _t('日志截取字数'), '显示单条日志时，截取日志内容字数。');
         $subMaxNum->input->setAttribute('class', 'mini');
         $form->addInput($subMaxNum);
 
@@ -110,12 +154,17 @@ class WeChatHelper_Plugin implements Typecho_Plugin_Interface
         /** APP_ID **/
         $appid = new Typecho_Widget_Helper_Form_Element_Text('WCH_appid', NULL, NULL,
             _t('APP_ID'), _t('需要管理菜单时填写，与开发模式服务器配置中填写一致。'));
+        $form->addInput($appid);
 
         /** APP_Secret **/
-        $form->addInput($appid);
         $appsecret = new Typecho_Widget_Helper_Form_Element_Text('WCH_appsecret', NULL, NULL,
             _t('APP_Secret'), _t('需要管理菜单时填写，与开发模式服务器配置中填写一致。'));
         $form->addInput($appsecret);
+
+        //禁用插件是否删除数据
+        $dropTable = new Typecho_Widget_Helper_Form_Element_Radio('dropTable',array('1' => _t('开启'), '0' => _t('关闭')),0,
+            _t('<span style="color:#B94A48">数据删除</span>'), _t('<span style="color:#B94A48">开启后，禁用插件会删除插件设置数据和数据表。</span>'));
+        $form->addInput($dropTable);
 
     }
